Refactor the packages structure of the operator
