+++ operator/app/src/main/java/org/keycloak/operator/controllers/StatusUpdater.java
@@ -1,4 +1,4 @@
-package org.keycloak.operator;
+package org.keycloak.operator.controllers;
 
 public interface StatusUpdater<T> {
