+++ operator/app/src/main/java/org/keycloak/operator/crds/v2alpha1/deployment/ValueOrSecret.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.operator.v2alpha1.crds;
+package org.keycloak.operator.crds.v2alpha1.deployment;
 
 import io.fabric8.kubernetes.api.model.SecretKeySelector;
